+++ hammerspoon/config/keymaps.lua
@@ -1,34 +1,34 @@
 Keymaps = {}
 
 -- Command mappers --
-Keymaps.Cmd = { 'cmd' }
-Keymaps.CmdShift = { 'cmd', 'shift' }
-Keymaps.CtrlAltCmd = { 'ctrl', 'alt', 'cmd' }
+Keymaps.CMD = { 'cmd' }
+Keymaps.CMD_REV = { 'cmd', 'shift' }
+Keymaps.LEADER_2 = { 'ctrl', 'alt', 'cmd' }
 
 -- Commands --
-hs.hotkey.bind(Keymaps.CtrlAltCmd, 'R', function()
+hs.hotkey.bind(Keymaps.LEADER_2, 'R', function()
   hs.reload()
 end)
 hs.alert.show('🔨 Hammerspoon reload')
 
 -- Leader Commands
-Keymaps.Leader = hs.hotkey.modal.new('alt', 'o')
+Keymaps.LEADER = hs.hotkey.modal.new('alt', 'o')
 
-function Keymaps.Leader:entered()
+function Keymaps.LEADER:entered()
   if hs.eventtap.isSecureInputEnabled() then
     hs.alert('⚠️ Secure Input is on. Hyper Mode commands might not work.')
   end
   print('triggered leader mode')
 end
 
-function Keymaps.Leader:exited()
+function Keymaps.LEADER:exited()
   print('exited from leader mode')
 end
-Keymaps.Leader:bind('', 'escape', function()
-  Keymaps.Leader:exit()
+Keymaps.LEADER:bind('', 'escape', function()
+  Keymaps.LEADER:exit()
 end)
 
-Keymaps.Leader:bind('alt', 'J', '🔗 google-chrome', ChromeWithProfile)
-Keymaps.Leader:bind('alt', 'K', '🔑 password-store', ChoosePassword)
+Keymaps.LEADER:bind('alt', 'J', '🔗 google-chrome', ChromeWithProfile)
+Keymaps.LEADER:bind('alt', 'K', '🔑 password-store', ChoosePassword)
 
 return Keymaps