+++ wezterm/config/init.lua
@@ -18,7 +18,7 @@ end
 function WezTermLoader:append(new_options)
   for k, v in pairs(new_options) do
     if self.options[k] ~= nil then
-      wezterm.log_warn('Duplicate config option detected: ', { old = self.options[k], new = new_options[k] })
+      wezterm.log_warn('[WezTerm] Duplicate config option detected: ',k, { old = self.options[k], new = new_options[k] })
       goto continue
     end
     self.options[k] = v